# contracts/openapi/v1.yaml
components:
  schemas:
    A2UIAction:
      additionalProperties: false
      description: Acción opaca ligada a un schema, una versión y un run concreto
        (§5.5).
      properties:
        action_id:
          maxLength: 67
          minLength: 5
          title: Action Id
          type: string
        expected_version:
          minimum: 1.0
          title: Expected Version
          type: integer
        input_schema_ref:
          maxLength: 300
          title: Input Schema Ref
          type: string
        label:
          maxLength: 120
          title: Label
          type: string
        requires_confirmation:
          default: true
          title: Requires Confirmation
          type: boolean
        tool_name:
          maxLength: 95
          title: Tool Name
          type: string
      required:
      - action_id
      - tool_name
      - input_schema_ref
      - expected_version
      - label
      title: A2UIAction
      type: object
...
    A2UIMessage:
      additionalProperties: false
      description: 'Unidad JSONL del protocolo: exactamente uno de los tres mensajes.'
      properties:
        createSurface:
          anyOf:
          - $ref: '#/components/schemas/CreateSurface'
          - type: 'null'
        updateComponents:
          anyOf:
          - $ref: '#/components/schemas/UpdateComponents'
          - type: 'null'
        updateDataModel:
          anyOf:
          - $ref: '#/components/schemas/UpdateDataModel'
          - type: 'null'
        version:
          default: v0.9.1
          title: Version
          type: string
      title: A2UIMessage
      type: object
    A2UISurface:
      additionalProperties: false
      description: 'Superficie completa: la secuencia ordenada de mensajes más sus
        acciones.'
      properties:
        actions:
          items:
            $ref: '#/components/schemas/A2UIAction'
          maxItems: 20
          title: Actions
          type: array
        catalog_id:
          description: URN inmutable del catálogo negociado, por ejemplo 'urn:nexo-ia:a2ui:catalog:citizen:v1'.
          maxLength: 200
          pattern: ^urn:nexo-ia:a2ui:catalog:[a-z0-9_-]+:v\d+$
          title: Catalog Id
          type: string
        channel:
          $ref: '#/components/schemas/Channel'
        messages:
          items:
            $ref: '#/components/schemas/A2UIMessage'
          maxItems: 100
          minItems: 1
          title: Messages
          type: array
        surface_id:
          maxLength: 68
          minLength: 6
          title: Surface Id
          type: string
      required:
      - surface_id
      - catalog_id
      - channel
      - messages
      title: A2UISurface
      type: object
    ActionResult:
      additionalProperties: false
      description: Resultado de ejecutar una acción confirmada (§5.1).
      properties:
        action_id:
          maxLength: 67
          minLength: 5
          title: Action Id
          type: string
        error:
          anyOf:
          - $ref: '#/components/schemas/NormalizedError'
          - type: 'null'
        idempotency_replayed:
          default: false
          title: Idempotency Replayed
          type: boolean
        status:
          $ref: '#/components/schemas/ActionStatus'
        tool_call_id:
          anyOf:
          - maxLength: 66
            minLength: 4
            type: string
          - type: 'null'
          title: Tool Call Id
        tool_result:
          anyOf:
          - $ref: '#/components/schemas/ToolResult'
          - type: 'null'
      required:
      - action_id
      - status
      title: ActionResult
      type: object
    ActionStatus:
      enum:
      - pending_confirmation
      - confirmed
      - succeeded
      - failed
      - partial
      - cancelled
      title: ActionStatus
      type: string
...
    Channel:
      enum:
      - web
      - whatsapp
      - voice
      title: Channel
      type: string
    ChannelFallback:
      additionalProperties: false
      description: 'Representación segura cuando la superficie no puede renderizarse
        (§5.5).


        Nunca queda vacío: si A2UI falla, el canal recibe texto plano equivalente
        en

        lugar de nada (`DIE-F1-106`).'
      properties:
        action_hint:
          anyOf:
          - maxLength: 300
            type: string
          - type: 'null'
          title: Action Hint
        channel:
          $ref: '#/components/schemas/Channel'
        numbered_items:
          items:
            type: string
          maxItems: 50
          title: Numbered Items
          type: array
        reason:
          maxLength: 200
          title: Reason
          type: string
        text:
          maxLength: 4000
          minLength: 1
          title: Text
          type: string
      required:
      - channel
      - reason
      - text
      title: ChannelFallback
      type: object
    ConfirmActionRequest:
      additionalProperties: false
      properties:
        consent:
          default: false
          title: Consent
          type: boolean
        expected_version:
          minimum: 1.0
          title: Expected Version
          type: integer
      required:
      - expected_version
      title: ConfirmActionRequest
      type: object
    Contradiction:
      additionalProperties: false
      description: Conflicto detectado entre hechos, fuentes o resultados de tools
        (§5.2).
      properties:
        conflicting_sources:
          items:
            maxLength: 67
            minLength: 5
            type: string
          maxItems: 20
          title: Conflicting Sources
          type: array
        conflicting_tool_calls:
          items:
            maxLength: 66
            minLength: 4
            type: string
          maxItems: 20
          title: Conflicting Tool Calls
          type: array
        contradiction_id:
          maxLength: 70
          minLength: 8
          title: Contradiction Id
          type: string
        explanation:
          maxLength: 1000
          title: Explanation
          type: string
        fact_ids:
          items:
            maxLength: 68
            minLength: 6
            type: string
          maxItems: 10
          minItems: 2
          title: Fact Ids
          type: array
        resolved_fact_id:
          anyOf:
          - maxLength: 68
            minLength: 6
            type: string
          - type: 'null'
          title: Resolved Fact Id
        rule:
          description: 'Regla de precedencia aplicada. Ejemplo: ''newer_source_wins''.'
          maxLength: 200
          title: Rule
          type: string
        severity:
          $ref: '#/components/schemas/ContradictionSeverity'
        status:
          $ref: '#/components/schemas/ContradictionStatus'
          default: detected
      required:
      - contradiction_id
      - fact_ids
      - severity
      - rule
      - explanation
      title: Contradiction
      type: object
...
    CreateSurface:
      additionalProperties: false
      properties:
        catalogId:
          description: URN inmutable del catálogo negociado, por ejemplo 'urn:nexo-ia:a2ui:catalog:citizen:v1'.
          maxLength: 200
          pattern: ^urn:nexo-ia:a2ui:catalog:[a-z0-9_-]+:v\d+$
          title: Catalogid
          type: string
        sendDataModel:
          default: false
          title: Senddatamodel
          type: boolean
        surfaceId:
          maxLength: 68
          minLength: 6
          title: Surfaceid
          type: string
      required:
      - surfaceId
      - catalogId
      title: CreateSurface
      type: object
    Domain:
      description: Los cinco namespaces iniciales (§2.3). Un dominio fuera de esta
        lista se rechaza.
      enum:
      - vehiculos
      - ayuntamiento_empresas
      - registro_civil
      - salud
      - ganaderia
      title: Domain
      type: string
    ErrorCode:
      description: 'Códigos de error estables (arquitectura §9.11).


        El consumidor decide por `code`, nunca por el texto del mensaje.'
      enum:
      - VALIDATION_ERROR
      - AUTHENTICATION_REQUIRED
      - PERMISSION_DENIED
      - RESOURCE_NOT_FOUND
      - APPOINTMENT_CONFLICT
      - VERSION_CONFLICT
      - ACTION_CONFIRMATION_REQUIRED
      - RATE_LIMITED
      - BUDGET_EXCEEDED
      - PROVIDER_ERROR
      - MODEL_UNAVAILABLE
      - MODEL_OUTPUT_INVALID
      - TOOL_TIMEOUT
      - TOOL_NOT_FOUND
      - RUN_TIMEOUT
      - RUN_CANCELLED
      - UNKNOWN_OUTCOME
      - CONFIGURATION_INVALID
      - CONTRACT_INVALID
      title: ErrorCode
      type: string
    ErrorDetail:
      additionalProperties: false
      description: Detalle accionable de un error, sin contenido sensible.
      properties:
        field:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          description: Ruta del campo que falló, en notación de puntos.
          title: Field
        reason:
          description: 'Motivo estable y corto, en snake_case. Ejemplo: ''conflict''.'
          maxLength: 200
          title: Reason
          type: string
      required:
      - reason
      title: ErrorDetail
      type: object
    Estimate:
      additionalProperties: false
      description: 'Resultado del estimador: pasos ordenados, costo total y trazabilidad.'
      properties:
        derived_from:
          items:
            maxLength: 68
            minLength: 6
            type: string
          maxItems: 200
          title: Derived From
          type: array
        domain:
          $ref: '#/components/schemas/Domain'
        steps:
          items:
            $ref: '#/components/schemas/EstimateStep'
          maxItems: 100
          title: Steps
          type: array
        total_cost:
          anyOf:
          - $ref: '#/components/schemas/Money'
          - type: 'null'
      required:
      - domain
      title: Estimate
      type: object
    EstimateStep:
      additionalProperties: false
      description: Un trámite dentro de la ruta calculada, con dependencias estables.
      properties:
        cost:
          anyOf:
          - $ref: '#/components/schemas/Money'
          - type: 'null'
        depends_on:
          items:
            pattern: ^[a-z][a-z0-9_]{1,62}$
            type: string
          maxItems: 20
          title: Depends On
          type: array
        derived_from:
          items:
            maxLength: 68
            minLength: 6
            type: string
          maxItems: 50
          minItems: 1
          title: Derived From
          type: array
        duration_days:
          anyOf:
          - maximum: 3650.0
            minimum: 0.0
            type: integer
          - type: 'null'
          title: Duration Days
        missing_documents:
          items:
            type: string
          maxItems: 30
          title: Missing Documents
          type: array
        step_id:
          pattern: ^[a-z][a-z0-9_]{1,62}$
          title: Step Id
          type: string
        title:
          maxLength: 200
          title: Title
          type: string
      required:
      - step_id
      - title
      - derived_from
      title: EstimateStep
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    JsonValue: {}
...
    Money:
      additionalProperties: false
      description: 'Monto en unidades menores más moneda (§2.3).


        Nunca se representa un monto en flotante ni se suma en un prompt: el

        estimador opera sobre `amount_minor` con código determinista.'
      properties:
        amount_minor:
          description: Monto en la unidad menor de la moneda; 125000 son 1,250.00
            MXN.
          title: Amount Minor
          type: integer
        currency:
          description: Código ISO 4217, por ejemplo 'MXN'.
          pattern: ^[A-Z]{3}$
          title: Currency
          type: string
      required:
      - amount_minor
      - currency
      title: Money
      type: object
    NormalizedError:
      additionalProperties: false
      description: 'Error normalizado que cruza cualquier frontera de Nexo IA.


        `outcome` es el campo que impide reintentos peligrosos: un `UNKNOWN`

        significa que no sabemos si la operación tuvo efecto, y por tanto nunca se

        reintenta de forma automática (`DIE-F0-035`).'
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        details:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 50
          title: Details
          type: array
        message:
          description: Mensaje para operadores. Nunca incluye prompts, secretos ni
            PII.
          maxLength: 500
          title: Message
          type: string
        outcome:
          $ref: '#/components/schemas/Outcome'
          default: known_failure
          description: Certeza sobre el efecto producido antes del fallo.
        retry_after_ms:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Retry After Ms
        retryable:
          description: Si la operación puede reintentarse tal cual, sin intervención.
          title: Retryable
          type: boolean
      required:
      - code
      - message
      - retryable
      title: NormalizedError
      type: object
    Outcome:
      description: 'Certeza sobre el efecto de una operación.


        `UNKNOWN` es el caso peligroso: un write con outcome desconocido nunca se

        reintenta automáticamente y produce `partial` (`DIE-F1-077`, `DIE-F1-081`).'
      enum:
      - known_success
      - known_failure
      - unknown
      title: Outcome
      type: string
    ProblemDetail:
      properties:
        code:
          title: Code
          type: string
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        errors:
          items:
            additionalProperties: true
            type: object
          title: Errors
          type: array
        retryable:
          default: false
          title: Retryable
          type: boolean
        status:
          title: Status
          type: integer
        title:
          title: Title
          type: string
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        type:
          default: about:blank
          title: Type
          type: string
      required:
      - title
      - status
      - code
      title: ProblemDetail
      type: object
...
    RunMetrics:
      additionalProperties: false
      description: Métricas obligatorias del run. El gate exige que el 100% las registre.
      properties:
        duration_ms:
          default: 0
          maximum: 3600000.0
          minimum: 0.0
          title: Duration Ms
          type: integer
        first_event_ms:
          anyOf:
          - maximum: 3600000.0
            minimum: 0.0
            type: integer
          - type: 'null'
          title: First Event Ms
        model_invocation_count:
          default: 0
          minimum: 0.0
          title: Model Invocation Count
          type: integer
        retrieval_count:
          default: 0
          minimum: 0.0
          title: Retrieval Count
          type: integer
        tool_call_count:
          default: 0
          minimum: 0.0
          title: Tool Call Count
          type: integer
        total_cost_usd:
          default: 0.0
          minimum: 0.0
          title: Total Cost Usd
          type: number
        total_input_tokens:
          default: 0
          minimum: 0.0
          title: Total Input Tokens
          type: integer
        total_output_tokens:
          default: 0
          minimum: 0.0
          title: Total Output Tokens
          type: integer
      title: RunMetrics
      type: object
    RunResult:
      additionalProperties: false
      description: 'Salida del núcleo (§5.1).


        Se construye desde `RunState` omitiendo todo lo interno: hechos candidatos

        sin verificar, nodos completados, reintentos e invocaciones de modelo no

        salen de la orquestación.'
      properties:
        answer:
          anyOf:
          - maxLength: 20000
            type: string
          - type: 'null'
          title: Answer
        available_actions:
          items:
            $ref: '#/components/schemas/A2UIAction'
          maxItems: 20
          title: Available Actions
          type: array
        error:
          anyOf:
          - $ref: '#/components/schemas/NormalizedError'
          - type: 'null'
        estimate:
          anyOf:
          - $ref: '#/components/schemas/Estimate'
          - type: 'null'
        fallback:
          anyOf:
          - $ref: '#/components/schemas/ChannelFallback'
          - type: 'null'
        metrics:
          $ref: '#/components/schemas/RunMetrics'
        run_id:
          maxLength: 67
          minLength: 5
          title: Run Id
          type: string
        sources:
          items:
            $ref: '#/components/schemas/SourceCitation'
          maxItems: 200
          title: Sources
          type: array
        status:
          $ref: '#/components/schemas/RunStatus'
        surface:
          anyOf:
          - $ref: '#/components/schemas/A2UISurface'
          - type: 'null'
        trace_id:
          maxLength: 69
          minLength: 7
          title: Trace Id
          type: string
        verified_facts:
          anyOf:
          - $ref: '#/components/schemas/VerifiedFacts'
          - type: 'null'
        warnings:
          items:
            type: string
          maxItems: 100
          title: Warnings
          type: array
      required:
      - run_id
      - trace_id
      - status
      title: RunResult
      type: object
    RunStatus:
      description: 'Estados de un run (arquitectura §9.11).


        `queued → planning → running → waiting_confirmation → running →

        succeeded | partial | failed`; `cancelled` puede ocurrir antes del final.'
      enum:
      - queued
      - planning
      - running
      - waiting_confirmation
      - succeeded
      - partial
      - failed
      - cancelled
      title: RunStatus
      type: string
    SourceCitation:
      additionalProperties: false
      description: 'Referencia puntual a un fragmento de una fuente (§5.2).


        Es inmutable a propósito: una citación es evidencia, y la evidencia no se

        edita después de recogerse.'
      properties:
        char_end:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Char End
        char_start:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          description: Inicio del tramo relevante dentro del fragmento.
          title: Char Start
        corpus_version:
          description: Versión de corpus vigente cuando se recuperó el fragmento.
          maxLength: 120
          title: Corpus Version
          type: string
        fragment_id:
          maxLength: 68
          minLength: 6
          title: Fragment Id
          type: string
        is_active:
          default: true
          description: Falso si la fuente fue sustituida o venció; bloquea claims
            críticos.
          title: Is Active
          type: boolean
        source_id:
          maxLength: 67
          minLength: 5
          title: Source Id
          type: string
        source_version:
          maxLength: 40
          title: Source Version
          type: string
        valid_from:
          format: date
          title: Valid From
          type: string
        valid_to:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Valid To
      required:
      - source_id
      - fragment_id
      - corpus_version
      - source_version
      - valid_from
      title: SourceCitation
      type: object
...
    ToolCallStatus:
      enum:
      - succeeded
      - failed
      - denied
      - timeout
      title: ToolCallStatus
      type: string
    ToolConfirmation:
      additionalProperties: false
      description: 'Comprobante verificable de una escritura.


        Sin folio, UUID o identificador equivalente no hay éxito: el resultado se

        degrada a `partial` (`DIE-F1-078`).'
      properties:
        identifier:
          description: Folio, UUID o identificador devuelto por el sistema destino.
          maxLength: 120
          minLength: 3
          title: Identifier
          type: string
        identifier_kind:
          default: folio
          maxLength: 40
          title: Identifier Kind
          type: string
        is_mock:
          default: true
          title: Is Mock
          type: boolean
        issued_at:
          format: date-time
          title: Issued At
          type: string
      required:
      - identifier
      - issued_at
      title: ToolConfirmation
      type: object
    ToolError:
      additionalProperties: false
      description: Error normalizado de una tool (§5.4).
      properties:
        error:
          $ref: '#/components/schemas/NormalizedError'
        provider:
          default: mock
          maxLength: 100
          title: Provider
          type: string
        safe_details:
          additionalProperties:
            $ref: '#/components/schemas/JsonValue'
          title: Safe Details
          type: object
      required:
      - error
      title: ToolError
      type: object
    ToolResult:
      additionalProperties: false
      description: Resultado de una invocación (§5.4).
      properties:
        confirmation:
          anyOf:
          - $ref: '#/components/schemas/ToolConfirmation'
          - type: 'null'
        data:
          additionalProperties:
            $ref: '#/components/schemas/JsonValue'
          title: Data
          type: object
        duration_ms:
          maximum: 3600000.0
          minimum: 0.0
          title: Duration Ms
          type: integer
        error:
          anyOf:
          - $ref: '#/components/schemas/ToolError'
          - type: 'null'
        idempotency_replayed:
          default: false
          title: Idempotency Replayed
          type: boolean
        is_mock:
          default: true
          title: Is Mock
          type: boolean
        name:
          description: Nombre `dominio.verbo_objeto`, por ejemplo 'vehiculos.reservar_cita'.
          maxLength: 95
          title: Name
          type: string
        provider:
          default: mock
          maxLength: 100
          title: Provider
          type: string
        status:
          $ref: '#/components/schemas/ToolCallStatus'
        tool_call_id:
          maxLength: 66
          minLength: 4
          title: Tool Call Id
          type: string
      required:
      - tool_call_id
      - name
      - status
      - duration_ms
      title: ToolResult
      type: object
    UpdateComponents:
      additionalProperties: false
      properties:
        components:
          items:
            $ref: '#/components/schemas/A2UIComponent'
          maxItems: 200
          minItems: 1
          title: Components
          type: array
        surfaceId:
          maxLength: 68
          minLength: 6
          title: Surfaceid
          type: string
      required:
      - surfaceId
      - components
      title: UpdateComponents
      type: object
    UpdateDataModel:
      additionalProperties: false
      properties:
        path:
          maxLength: 300
          pattern: ^/[A-Za-z0-9_/-]*$
          title: Path
          type: string
        surfaceId:
          maxLength: 68
          minLength: 6
          title: Surfaceid
          type: string
        value:
          $ref: '#/components/schemas/JsonValue'
      required:
      - surfaceId
      - path
      - value
      title: UpdateDataModel
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    VerifiedFact:
      additionalProperties: false
      description: 'Hecho tras pasar por el verificador (§5.2).


        Invariante central: si es crítico y fue aceptado, tiene al menos una

        citación activa y es apto para escritura solo si además está aceptado.'
      properties:
        category:
          $ref: '#/components/schemas/FactCategory'
        citations:
          items:
            $ref: '#/components/schemas/SourceCitation'
          maxItems: 20
          title: Citations
          type: array
        claim:
          maxLength: 1000
          title: Claim
          type: string
        confidence:
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          type: number
        depends_on:
          items:
            maxLength: 68
            minLength: 6
            type: string
          maxItems: 50
          title: Depends On
          type: array
        domain:
          $ref: '#/components/schemas/Domain'
        fact_id:
          maxLength: 68
          minLength: 6
          title: Fact Id
          type: string
        reason:
          description: 'Motivo estable de la decisión. Ejemplo: ''source_expired''.'
          maxLength: 300
          title: Reason
          type: string
        supporting_tool_call_id:
          anyOf:
          - maxLength: 66
            minLength: 4
            type: string
          - type: 'null'
          description: Invocación de tool que respalda el hecho cuando su evidencia
            no es documental. Es la otra forma admisible de fundamentar un claim crítico.
          title: Supporting Tool Call Id
        value:
          $ref: '#/components/schemas/FactValue'
        verification:
          $ref: '#/components/schemas/VerificationStatus'
        write_eligible:
          default: false
          title: Write Eligible
          type: boolean
      required:
      - fact_id
      - claim
      - value
      - category
      - domain
      - verification
      - reason
      - confidence
      title: VerifiedFact
      type: object
    VerifiedFacts:
      additionalProperties: false
      description: 'Snapshot inmutable que reciben estimación final, A2UI y redactor
        (§5.2).


        El redactor solo ve esto: no tiene puertos de RAG ni de MCP, así que este

        snapshot es literalmente el universo de hechos del que puede hablar

        (`DIE-F1-094`).'
      properties:
        contradictions:
          default: []
          items:
            $ref: '#/components/schemas/Contradiction'
          maxItems: 100
          title: Contradictions
          type: array
        created_at:
          format: date-time
          title: Created At
          type: string
        facts:
          items:
            $ref: '#/components/schemas/VerifiedFact'
          maxItems: 500
          title: Facts
          type: array
        snapshot_id:
          maxLength: 120
          title: Snapshot Id
          type: string
      required:
      - snapshot_id
      - created_at
      - facts
      title: VerifiedFacts
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: 'API del backend (Dani). Convenciones: JSON snake_case, UTC, IDs opacos
    (`conv_`, `run_`, `act_`, `apt_`), dinero en unidades menores. Los errores usan
    **Problem Details** (`type/code/status/retryable`); el frontend nunca parsea texto.
    Toda respuesta propaga `X-Trace-Id`. Auth = Supabase Auth (JWT bearer validado
    por JWKS).'
  summary: Gateway HTTP/SSE/webhooks del asistente institucional nexo-ai.
  title: nexo-ai API
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/v1/actions/{action_id}/confirm:
    post:
      operationId: confirm_action_api_v1_actions__action_id__confirm_post
      parameters:
      - in: path
        name: action_id
        required: true
        schema:
          title: Action Id
          type: string
      - in: header
        name: Idempotency-Key
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmActionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Error de validación (VALIDATION_ERROR)
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Autenticación requerida (AUTHENTICATION_REQUIRED)
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Permiso denegado (PERMISSION_DENIED)
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Conflicto (APPOINTMENT_CONFLICT / VERSION_CONFLICT)
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Confirmación requerida (ACTION_CONFIRMATION_REQUIRED)
      security:
      - HTTPBearer: []
      summary: Confirmar acción (idempotente)
      tags:
      - actions
...
  /api/v1/runs/{run_id}:
    get:
      operationId: get_run_api_v1_runs__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResult'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Autenticación requerida (AUTHENTICATION_REQUIRED)
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Recurso no encontrado (RESOURCE_NOT_FOUND)
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Snapshot de un run
      tags:
      - runs
  /api/v1/runs/{run_id}/events:
    get:
      operationId: run_events_api_v1_runs__run_id__events_get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: query
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - in: header
        name: Last-Event-ID
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Last-Event-Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
            text/event-stream: {}
          description: Stream de RunEvent (text/event-stream), reanudable por Last-Event-ID.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Autenticación requerida (AUTHENTICATION_REQUIRED)
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Recurso no encontrado (RESOURCE_NOT_FOUND)
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Stream SSE de eventos del run
      tags:
      - runs
...